Rename modules and files.
